requirements.txt
certifi==2022.12.7
charset-normalizer==2.0.12
idna==3.4
PyYAML==6.0
requests==2.26.0
urllib3==1.26.13
